user_profile_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
cryptography